SEQUOIADBMAINSTREAM-3992 【BUG】【new】【story】【S3对象管理】设置不同版本控制状态添加对象,对象删除后,其lob数据有残留 删除历史元数据中版本号为null的记录时，同时删除lob1的对象数据
git-svn-id: http://192.168.20.11/sequoiadb/trunk@38054 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sequoias3/src/main/java/com/sequoias3/service/impl/ObjectServiceImpl.java
@@ -1222,11 +1222,17 @@ public class ObjectServiceImpl implements ObjectService {
                                 metaResult, 1, true);
                         metaDao.updateMeta(connection, metaCsName, metaClName, bucketId,
                                 objectName, null, objectMeta);
+                        ObjectMeta nullMeta = null;
                         if (VersioningStatusType.SUSPENDED == versioningStatusType) {
-                            metaDao.removeMeta(connection, metaCsName, metaHisClName, bucketId,
-                                    objectName, null, true);
+                            nullMeta = metaDao.queryForUpdate(connection, metaCsName, metaHisClName,
+                                    bucketId, objectName, null, true);
+                            if (null != nullMeta) {
+                                metaDao.removeMeta(connection, metaCsName, metaHisClName, bucketId,
+                                        objectName, null, true);
+                            }
                         }
                         transaction.commit(connection);
+                        deleteObjectLob(nullMeta);
                     }
                 }
                 return;
